Create User for UID 1001
This commit creates a user to match UID 1001, which is the default user
for these containers. This will fix some of the quirky behavior that
occurs when running as a UID with no backing user.

This change attempts to keep the behavior as close to what was happening
previously, in order to prevent breaking anyone's build who relies on
the existing behavior. This _should_ be a non-breaking change, but no
guarantees.